+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 # unreleased
 
+# 8.3.0
+
+- fix `FlowScripts` type declaration
+- throw `FlowResourceError` if failing to get script resource content
+
 # 8.2.2
 
 - bump [0dep/piso@2](https://www.npmjs.com/package/@0dep/piso)